% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{test_cohort}
\alias{test_cohort}
\title{Testing Cohort for Testing}
\format{
a data.frame with 100,000 columns and 6 variables
\describe{
\item{age}{Age of the patent}
\item{sex}{Example biological sex}
\item{race}{Example Race/Ethnicity}
\item{vaccinated}{Binary indicator if the patient was vaccinated}
\item{infected}{Binary indicator if the patient was infected}
\item{cohort_id}{An identifier for the cohort}
}
}
\usage{
test_cohort
}
\description{
Testing Cohort for Testing
}
\keyword{datasets}
