% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecalc-package.R
\docType{package}
\name{vecalc-package}
\alias{vecalc}
\alias{vecalc-package}
\title{vecalc: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Michael DeWitt \email{me.dewitt.jr@gmail.com} (\href{https://orcid.org/0000-0001-8940-1967}{ORCID})

}
\keyword{internal}
