% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_ve.R
\name{fit_ve}
\alias{fit_ve}
\title{Fit Vaccine Efficiency}
\usage{
fit_ve(ve_data, stan_opts = list())
}
\arguments{
\item{ve_data}{an object of class vedata}

\item{stan_opts}{a list, any additional parameters to pass to
Stan.}
}
\description{
A function for fitting vaccine efficiency
}
