% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_ve.R
\name{prepare_ve_data}
\alias{prepare_ve_data}
\title{Prepare Data for Matching}
\usage{
prepare_ve_data(form, data, method = "cem")
}
\arguments{
\item{form}{the formal to be used in the matching algorithm}

\item{data}{a data.frame with the data objects to be used}

\item{method}{a string indicating the matching algorithm to be
used.}
}
\description{
This function returns objects that be used for calculating
Vaccine Efficiency
}
